Improve the error message from process::validate
When we fail to look up the container command in $PATH, the error
message would be " not found in $PATH: No such file or directory" if the
config.process.args was empty. This adds quotation marks to make it a
little clearer that we are failing to find a program called "" in $PATH.

This is similar to the corresponding error message generated by crun on
Linux.